feat: :sparkles: Introducing the rework for the post page
